requirements.txt
discord.py
python-dotenv
stripe
paypalrestsdk
aiosqlite
fastapi
uvicorn
aiohttp
